A lot

+++ src/server/databases.zig
@@ -7,6 +7,8 @@ const auth = @import("auth.zig");
 
 const DATA_DIR = "databases/";
 const DB_EXTENSION = ".sqlite3";
+/// Characters forbidden from DB names, to prevent arbitrary filesystem access
+const FORBIDDEN_CHARS = "/.";
 
 pub fn manifest(app: *root.App, req: *httpz.Request, res: *httpz.Response) !void {
     const username = try auth.authenticate(app, req, res, .header) orelse {
@@ -18,11 +20,7 @@ pub fn manifest(app: *root.App, req: *httpz.Request, res: *httpz.Response) !void
         timestamp: i64,
     };
 
-    const stmt = try (try app.db_conn.prepare(
-        res.arena,
-        "SELECT name, UNIX_TIMESTAMP(modified) FROM db_manifests WHERE username = ?",
-    )).expect(.stmt);
-    const rows = try (try app.db_conn.executeRows(&stmt, .{username}))
+    const rows = try (try app.db_conn.executeRows(&app.stmts.sel_manifests, .{username}))
         .expect(.rows);
     var iter = rows.iter();
 
@@ -46,6 +44,9 @@ pub fn download(app: *root.App, req: *httpz.Request, res: *httpz.Response) !void
         res.body = "Missing database name";
         return;
     };
+    if (!checkName(db_name, res)) {
+        return;
+    }
 
     const db_path = try std.mem.concat(res.arena, u8, &.{ DATA_DIR, username, "/", db_name, DB_EXTENSION });
 
@@ -76,6 +77,9 @@ pub fn upload(app: *root.App, req: *httpz.Request, res: *httpz.Response) !void {
         res.body = "Missing database name";
         return;
     };
+    if (!checkName(db_name, res)) {
+        return;
+    }
 
     const body = req.body() orelse {
         res.status = 400;
@@ -87,12 +91,7 @@ pub fn upload(app: *root.App, req: *httpz.Request, res: *httpz.Response) !void {
 
     try std.compress.gzip.decompress(fbs.reader(), decompressed.writer());
 
-    const stmt = try (try app.db_conn.prepare(
-        res.arena,
-        "INSERT INTO db_manifests VALUES (?, ?, NOW()) ON DUPLICATE KEY UPDATE modified = NOW()",
-    )).expect(.stmt);
-
-    _ = try (try app.db_conn.execute(&stmt, .{ username, db_name })).expect(.ok);
+    _ = try (try app.db_conn.execute(&app.stmts.ins_or_upd_manifest, .{ username, db_name })).expect(.ok);
 
     const user_dir = try std.mem.concat(res.arena, u8, &.{ DATA_DIR, username });
     try std.fs.cwd().makePath(user_dir);
@@ -113,15 +112,25 @@ pub fn delete(app: *root.App, req: *httpz.Request, res: *httpz.Response) !void {
         res.body = "Missing database name";
         return;
     };
+    if (!checkName(db_name, res)) {
+        return;
+    }
 
-    const stmt = try (try app.db_conn.prepare(
-        res.arena,
-        "DELETE FROM db_manifests WHERE username = ? AND name = ?",
-    )).expect(.stmt);
-
-    _ = try (try app.db_conn.execute(&stmt, .{ username, db_name })).expect(.ok);
+    _ = try (try app.db_conn.execute(&app.stmts.del_manifest, .{ username, db_name })).expect(.ok);
 
     const db_path = try std.mem.concat(res.arena, u8, &.{ DATA_DIR, username, "/", db_name, DB_EXTENSION });
 
     std.fs.cwd().deleteFile(db_path) catch {};
+}
+
+fn checkName(name: []const u8, res: *httpz.Response) bool {
+    for (FORBIDDEN_CHARS) |chr| {
+        if (std.mem.containsAtLeast(u8, name, 1, &.{chr})) {
+            res.status = 400;
+            res.body = "Invalid characters in database name";
+            return false;
+        }
+    }
+
+    return true;
 }
